Fix datatype mismatch

+++ ProjektBack/src/main/resources/schema.sql
@@ -255,7 +255,7 @@ CREATE TABLE  Klient
 
 CREATE TABLE  Laadimispunkti_kategooria_omamine 
 (
-	 Laadimispunkti_kood  smallint NOT NULL,
+	 Laadimispunkti_kood  bigint NOT NULL,
      laadimispunkti_kategooria_kood  smallint NOT NULL,
 	CONSTRAINT  PK_Laadimispunkti_kategooria_omamine  PRIMARY KEY ( Laadimispunkti_kood , laadimispunkti_kategooria_kood ),
 	CONSTRAINT FK_laadimispunkti_kood FOREIGN KEY (Laadimispunkti_kood) REFERENCES laadimispunkt(laadimispunkti_kood) ON UPDATE CASCADE ,
